:truck: rename s/server/api
