+++ api/compose.dev.yml
@@ -1,9 +1,9 @@
 services:
-  server:
+  api:
     build:
       context: .
       dockerfile: Dockerfile.dev
-    container_name: server
+    container_name: api
     volumes:
       - ./:/app
       - ~/.aws:/root/.aws
