+++ api/go.mod
@@ -1,4 +1,4 @@
-module server
+module api
 
 go 1.22
